Adds lliascript option to ignore existing buffer or bus.
Previously an attempt to add a buffer or bus with an existing name
intentionally caused an error.  This was fine interactively but becomes a
hassle when executing an external batch file.

Now abus, cbus, buffer and related commands have the keyword option :nodup
If :nodup is 1 the behavior is as before, attempting to add a bus/buffer
with an existing name causes an error.   If :nodub is 0, no error is
generated and the script simply skips the command.