Adds synth GroupWindow class.
A GroupWindow is a Tk Toplevel window for holding active synth
editors in a Tk Notebook widget.

Prior to GroupWindow each synth editor had it's own Toplevel
which was both cluttered and confusing.

Additional GroupWindows may be created from the main application
window.  As new synths are added they are place in the newest
GroupWindow.