Integrates new Bus objects,
Made modifications to proxy, parser and add bus dialogs to support
new AudioBus and ControlBus objects.

Previously proxy maintained a list of active buses as a simple tuple
with bus name and channel count.   This did not allow tracking of
of how buses are being used.  The new bus objects contains fields
for source and sink destinations.