Replaces depreciated TkBusAndBufferEditor with TkBusConnectionEditor.
TkBusAndBufferEditor was a static list of bus assignments, and there was no
way to alter bus connections after a synth was created.

The new TkBusConnectionEditor allows bus assignments to be made -after- a
synth is already in existence.

Caveat:
There are still restrictions imposed by the
order in which synths are created.  Effect synths are chained to in
-reverse- of the order they are created.   TkBusConnectionEditor does not
check that synths are connection in a valid way.  If you attempt to
connect two are more effects in an invalid order, there probably will not be
audio output.