Automated synth identification numbers.
Each SynthProxy has an id_ field.  Previously the id_ had to be unique for
any given synth type.  The new behavior is for the id_ to be unique
across -ALL- synths.  These id numbers are now automatically generated
starting with 0 and incrementing with each new synth.

It is now trivial to determine the relative creation times between two
synths.  This information is useful primarily for effect synths where the
order in which they are created restricts which synth may serve as
inputs/outputs to other synths.