Brings compose feature up to date.
Compose creates a python script which may be used to recreate the state of
Llia.  The result of executing the script is a functionally equivalent graph
of Llia when the script was created. The resulting graph is isomorphic to
the original but at this time does not preserve the visual appearance
(maybe later).   On reconstruction nodes are placed randomly on the canvas,
some manual manipulation may be required if you want it to look nice.