BUGFIX  Bug 0009 closed.
Creating new Group windows no longer causes an error.
The fix was to provide a separate set of serial numbers to keep
track of the creation order for all synths and groups.

This series of numbers is separate from the synth serial number
which makes up part of the synth_id sid.