Additional code for graph auto align. Extends commit a409526.
Clicking the Align button places all tokens in rows an columns. Synth
tokens are sorted right to left by serial number illustrating the order in
which they may be connected.

An attempt is made to place audio and control buses in proximity to synth
outputs they are connected to, however a bus may be connected to several
synth outputs.

All unused buses are 'parked' on the lower right side.

The graph can become quite cluttered and some thought needs to dedicated to
cleaner auto align algorithms.  One idea to reduce clutter is to add
optional user filters to selectively display or hide various elements.