Adds new PolyN keymode.
Unlike Poly1, which allocates synths on the fly, PolyN creates a
fixed number of synths ahead of time. This PolyN a fixed resource load
but at the cost of (nearly) unlimited polyphony.

PolyN is to become the new default keymode.

At this point only Orgn has been adapted to use PolyN