BUG FIX 0015 KeyError on MSB updates.
Added fallback method to be used if MSB.value fails.
The fallback checks each assigned value against the new_value and selects
the one which is closest.  This only works for numeric values and is
highly inefficient.

However the most likely cause for using the fallback is due to float
rounding errors and typically only a few values need to be checked.