Adds control widget creation methods to TkSubEditor.
Most current sub classes of TkSubEditor define similar functions
which makes for a lot of code duplication.   These new methods
are to eventually replace the previous ones.

NOTE: A few very early editors do not use as Tk Canvas as their
background.  These editors will have to be modified to make use
of the new methods.