Updates serialization for Program, Performance and ProgramBank.
The previous serialization method used tagged values for each and every
program parameter and each program was saved explicitly with the bank.
This often resulted in large data files with sizes on the order of a few
megs.

The new behavior uses positional values in place of tags.  Additionally if
two or more sequential programs in a bank are identical, only data for the
first program is saved. A special marker is used to indicate repeated
programs.  This saves considerable space, particularly with initialized
banks with only a few custom programs.  The amount of savings is dependent
on the make up of the bank. Using the default Algo bank with 16 defined
programs, the new bank file is approximately 4% the previous value
(116k vs 2.8 meg).   On the minus side bank load times have increased.