Adds new Scene feature.
A scene is an extension of the compose feature.  Like compose a scene
captures the current configuration of Llia for later recall.  The primary
difference between a scene and compose is that a scene combines all
elements into a single unified file.

Motivation:

While using Llia in a studio session the compose feature was used to save
the current configuration.  This worked but in addition to the compose
file there were half a dozen or so bank data files.   Any future
attempt to use compose to reconstruct Llia would be predicated on locating
these files.  If any bank file was moved compose could not recreate
the corresponding synth configuration.

Scene was developed to address this issue.  A scene combines the compose
script along with all bank data into a single file.  There were a few
changes made in preparation for scenes.

1) Bank data file structure was made more efficient.
2) A Progress bar was added to the GUI.

Still to do:

1) The scene filename is hard coded, a file dialog needs to be added.
2) The compose feature is still available to the user under the file menu.
   This will probably go away, although compose itself will remain as it
   is required by scene.